Replaced all checks for python with python2
This makes it so, if both Python 2 and 3 are present on the system,
Python 2 will get used. Executing just 'python' will run either the more
recently installed version or the one with a higher version number.
'python2' guarantees that Python 2 will get used if present

Tested with ActivePython 2.7. **PLEASE CHECK WITH NORMAL PYTHON 2.7**