Enable premake linkgroups when building Linux applications
The linkgroups option is a convenience option that makes it
possible to specify all the libraries required to build an
application (console or GUI) in any order. It also prevents
circular dependency issues between libraries.

Basically, it surrounds the list of libraries with the
start-group/end-group options when calling the linker.