Adjust resolve texture sizes to match renders that filled them
When a copy command is being processed, the only information available about
the size of the destination texture is its tile dimensions, which are
frequently larger than the actual pixel size. Keep track of the window rects
used during renders and then try and match a previously used window rect to
the texture tile size during the resolve process to correctly crop the texture.