Texture region splitting
Introduce the concept of texture regions. Each texture starts with a base
region which covers all of its content, and can provide smaller subregions
given an offset and size (populated by allocating a new VkImage and blitting
from the base region). Subregion allocations and content are cached and
invalidated when the base region is written. Writing is only allowed to the
base region at this time.