[Kernel] Add more XConfig settings & --xconfig_initial_setup flag
This just adds some extra XConfig settings that the dashboard tries to retrieve.

The XCONFIG_USER_RETAIL_FLAGS setting gets used to decide whether to show the initial setup/OOBE wizard or not.
Since the OOBE currently isn't usable in Xenia (mostly due to missing XamProfile* exports, afaik), this setting is by default set to skip it.
But if that isn't desired the --xconfig_initial_setup flag can be used to un-set it.