Add magic division number computation for constant unsigned 64 bit divisions and 32 bit divisions (add flag is not handled right now, divs requiring that instead do the full x86 div)
For 32 and 64 bit values, constant multiplications by 3, 5, and 9 will
now be done with lea.

Integer adds of all widths now check to see if they can be inc instead
Integer subs of all widths now check to see if they can be dec instead

Added fast path for VECTOR_SHL::EmitInt8 that only requires three
instructions if src2 is constant and every value is the same, which is
universal on Halo: Reach at least.