[x64] Move all sequences in same compilation unit
Currently, sequences are in separate compilation units, but their
constructors all share the same unordered_map, which is in one of the
compilation unit.

Due to that, when compiling using GCC with LTO enabled, the constructor of
some units is called before the constructor of the unit with the map, also
called 'Static Initialization Order Fiasco'
(see: https://en.cppreference.com/w/cpp/language/siof). Thus, the
map is not initialized, which causes a division-by-zero error (SIGFPE).

Fix this by converting the files that contain specific sequences to be
includable in the main file. This way all sequences are in the same
compilation unit and guarantees the map is initialized.